created hotelDetailViewModel and HotelDetailViewModelFactory

+++ app/src/main/java/com/example/myhotels/domain/model/HotelDetailItem.kt
@@ -4,7 +4,7 @@ data class HotelDetailItem(
     val id: Int,
     val name: String,
     val address: String,
-    val stars: Double,
+    val stars: Int,
     val distance: Double,
     val suitesAvailability: String,
     val image: String,
